# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Clone API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl: string | null;
  createdAt: string;
}

export interface LoginBody {
  username: string;
  displayName: string;
}

export interface Channel {
  id: number;
  name: string;
  createdAt: string;
}

export interface CreateChannelBody {
  name: string;
}

export interface Message {
  id: number;
  content: string;
  channelId: number;
  userId: number;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl: string | null;
  createdAt: string;
}

export interface SendMessageBody {
  content: string;
}