# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Clone API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  displayName: zod.string(),
  avatarUrl: zod.string().nullable(),
  createdAt: zod.string(),
});

/**
 * @summary Login with username
 */
export const LoginBody = zod.object({
  username: zod.string(),
  displayName: zod.string(),
});

export const LoginResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  displayName: zod.string(),
  avatarUrl: zod.string().nullable(),
  createdAt: zod.string(),
});

/**
 * @summary List all channels
 */
export const ListChannelsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  createdAt: zod.string(),
});
export const ListChannelsResponse = zod.array(ListChannelsResponseItem);

/**
 * @summary Create a channel
 */
export const CreateChannelBody = zod.object({
  name: zod.string(),
});

/**
 * @summary List messages in a channel
 */
export const ListMessagesParams = zod.object({
  channelId: zod.coerce.number(),
});

export const ListMessagesResponseItem = zod.object({
  id: zod.number(),
  content: zod.string(),
  channelId: zod.number(),
  userId: zod.number(),
  username: zod.string(),
  displayName: zod.string(),
  avatarUrl: zod.string().nullable(),
  createdAt: zod.string(),
});
export const ListMessagesResponse = zod.array(ListMessagesResponseItem);

/**
 * @summary Send a message to a channel
 */
export const SendMessageParams = zod.object({
  channelId: zod.coerce.number(),
});

export const SendMessageBody = zod.object({
  content: zod.string(),
});